fix(server): fix route-extractor for two-level indirection (t/2749)
routes/registerAll.ts → routes/all.ts so extractRoutes resolveClusterFile
finds it (registerAllRoutes → cluster "All" → routes/all.ts). Also add a
same-dir fallback in resolveClusterFile so routes within all.ts (which call
registerMetaRoutes etc. as siblings) resolve correctly on the Linux CI
filesystem. Without the fallback extractRoutes got 0 routes and the
routeTable snapshot + publicShare route-scope test failed.
